First test performed.

+++ simple_tf/cign/tree.py
@@ -302,7 +302,7 @@ class TreeNetwork:
         gumbel_sample = -1.0 * tf.log(-1.0 * tf.log(uniform_sample))
         log_probs = tf.log(probs)
         gumbel_max = gumbel_sample + log_probs
-        selected_indices = tf.argmax(gumbel_max, axis=1)
+        selected_indices = tf.cast(tf.argmax(gumbel_max, axis=1), tf.int32)
         return selected_indices
 
     def calculate_accuracy(self, calculation_type, sess, dataset, dataset_type, run_id, iteration):
